requirements.txt
azure-cognitiveservices-speech==1.22.0
spacy==3.1.3
scispacy==0.4.0
textblob==0.15.3
